fix: address AI review — defensive URL, Vite crash handling, dedup utils, env naming
- Add fallback base to `new URL(c.req.url, "http://localhost")` for safety
- Handle Vite crash in dev-website.mjs (tear down Astro if Vite exits)
- Extract shared port helpers to `scripts/dev-utils.mjs`
- Rename `VITE_VIEWER_URL` → `DEV_VIEWER_URL` to avoid Vite env convention confusion

via [Happy](https://happy.engineering)
