+++ scripts/dev.mjs
@@ -8,42 +8,10 @@
  *   node scripts/dev.mjs -d         # dashboard mode (passes -d to CLI)
  */
 import { spawn } from "node:child_process";
-import { createServer } from "node:net";
+import { findFreePort } from "./dev-utils.mjs";
 
 const VITE_PREFERRED = 5173;
 const API_PREFERRED = 13456;
-
-/**
- * Check if a port is free on BOTH IPv4 and IPv6 (macOS Vite listens on ::1).
- * Returns true if the port is available.
- */
-function isPortFree(port) {
-  return new Promise((resolve) => {
-    // Check IPv4 first
-    const s4 = createServer();
-    s4.unref();
-    s4.on("error", () => resolve(false));
-    s4.listen(port, "127.0.0.1", () => {
-      s4.close(() => {
-        // Then check IPv6
-        const s6 = createServer();
-        s6.unref();
-        s6.on("error", () => resolve(false));
-        s6.listen(port, "::1", () => {
-          s6.close(() => resolve(true));
-        });
-      });
-    });
-  });
-}
-
-/** Find a free port starting from `preferred`, incrementing on conflict. */
-async function findFreePort(preferred) {
-  for (let port = preferred; port < preferred + 100; port++) {
-    if (await isPortFree(port)) return port;
-  }
-  throw new Error(`No free port found in range ${preferred}-${preferred + 99}`);
-}
 
 const dashboardMode = process.argv.includes("-d");
